# unpack: accept fflag as an option