--- a/lib/unpack.js
+++ b/lib/unpack.js
@@ -159,6 +159,7 @@ class Unpack extends Parser {
     this.umask = typeof opt.umask === 'number' ? opt.umask : this.processUmask
     // default mode for dirs created as parents
     this.dmode = opt.dmode || (0o0777 & (~this.umask))
+    this.fflag = opt.fflag || 'w'
     this.fmode = opt.fmode || (0o0666 & (~this.umask))
     this.on('entry', entry => this[ONENTRY](entry))
   }
@@ -294,6 +295,7 @@ class Unpack extends Parser {
   [FILE] (entry) {
     const mode = entry.mode & 0o7777 || this.fmode
     const stream = new fsm.WriteStream(entry.absolute, {
+      flags: this.fflag,
       mode: mode,
       autoClose: false
     })
@@ -515,7 +517,7 @@ class UnpackSync extends Unpack {
     let stream
     let fd
     try {
-      fd = fs.openSync(entry.absolute, 'w', mode)
+      fd = fs.openSync(entry.absolute, this.fflag, mode)
     } catch (er) {
       return oner(er)
     }