This allows the WriteStream flags to be used when writing files to be specified as an option.

This allows programs using node-tar to take advantage of writing files using a file mapping, which can be significantly faster on Windows. Support landed in libuv in https://github.com/libuv/libuv/pull/2295 and in Node.js in https://github.com/nodejs/node/pull/29260.

For the included benchmarks, execution time drops from 6.6s to 2.8s for async and from 15.8s to 3.4s for sync on my local machine. This depends heavily on the machine being used: can range from negligible improvements to even better than the above.