@joaocgreis 2019-09-17
@isaacs updated, thanks for reviewing!

UV_FS_O_FILEMAP should not be enabled by default in tar, because it is slower for large files (>1MB).

It should however be enabled by default in npm where most files are small: https://github.com/npm/pacote/pull/8