@isaacs 2019-09-18
> UV_FS_O_FILEMAP should not be enabled by default in tar, because it is slower for large files (>1MB).

The file size is known in advance when extracting a tar entry, so it seems like we _should_ default it for cases where we know it's going to be a benefit, no?