@joaocgreis 2019-09-21
@isaacs if node-tar can accept a solution that doesn't use an option, then sure! Here it is: https://github.com/npm/node-tar/pull/230