@joaocgreis 2019-09-21
Closing in favor of https://github.com/npm/node-tar/pull/230